// server/src/types/file-upload-types.ts
export interface FileInfo {
  fieldname: string;
  filename: string;
  encoding: string;
  mimetype: string;
  size: number;
  destination?: string;
  path?: string;
  originalFilename?: string;
  publicUrl?: string;
  cloudinaryId?: string;
  resourceType?: string;
}

export interface UploadOptions {
  maxFileSize?: number;
  allowedMimeTypes?: string[];
  destination?: string;
  pathStructure?: string;
  cloudinaryFolder?: string;
  useCloudinary?: boolean;
}

export interface UploadResult {
  files: FileInfo[];
  fields: Record<string, any>;
}

// export interface FileInfo {
//   fieldname: string;
//   filename: string;
//   originalFilename: string;
//   encoding: string;
//   mimetype: string;
//   size: number;
//   destination: string;
//   path: string;
//   publicUrl: string;
// }

// export interface UploadOptions {
//   maxFileSize?: number;
//   allowedMimeTypes?: string[];
//   metadata?: Record<string, string>;
//   destination?: string;
//   pathStructure?: string;
//   useOriginalFilename?: boolean;
// }

// export interface UploadResult {
//   files: FileInfo[];
//   fields: Record<string, any>;
// }